+++ src/containers/AddFunds/AddFunds.types.ts
@@ -1,0 +1,4 @@
+export interface AddFundsProps {
+	visible: boolean;
+	onDismiss: () => void;
+}
