+++ src/containers/AddFunds/ChooseQuantityModal/ChooseQuantityModal.types.ts
@@ -1,0 +1,9 @@
+import { ICoin } from '@helpers/coins';
+
+export interface ChooseQuantityModalProps {
+	coin: ICoin;
+	amount: number | undefined;
+	setPresetAmount: Function;
+	enableCustomAmount: () => void;
+	onPurchase: () => void;
+}
